# discussion: optimize for very short arrays? 