I know that there is a bug still in there, but before i work further on this, I wanted to show this as a discussion point.

I believe that most arrays are very short. Like 0 to 10 entries. Of course we can have like Arrays with 1e4 and more entries. We can add a fast path for short arrays. I benched this approach and up to 3 elements it is much faster than using the for loop. 

this branch wth minSample 100:

```sh
node benchmark/bench-array.js
empty array................... x 240,717,906 ops/sec ±0.15% (193 runs sampled)
array with one number value... x 122,848,208 ops/sec ±0.58% (188 runs sampled)
array with two number values.. x 52,701,578 ops/sec ±0.12% (195 runs sampled)
array with three number values x 28,192,651 ops/sec ±0.11% (195 runs sampled)
array with three four values.. x 15,970,103 ops/sec ±0.14% (194 runs sampled)
```

same benchmark against master:

```sh
 node benchmark/bench-array.js
empty array................... x 243,649,239 ops/sec ±0.10% (189 runs sampled)
array with one number value... x 87,170,806 ops/sec ±0.35% (191 runs sampled)
array with two number values.. x 35,999,960 ops/sec ±0.65% (193 runs sampled)
array with three number values x 21,870,022 ops/sec ±0.42% (192 runs sampled)
array with three four values.. x 16,138,385 ops/sec ±0.14% (195 runs sampled)
```

So there are maybe some perf gains possible. If you give green light, I would work further on it.

<!--
Thank you for your pull request. Please provide a description above and review
the requirements below.

Bug fixes and new features should include tests and possibly benchmarks.

Contributors guide: https://github.com/fastify/fastify/blob/master/CONTRIBUTING.md

By making a contribution to this project, I certify that:

* (a) The contribution was created in whole or in part by me and I
  have the right to submit it under the open source license
  indicated in the file; or

* (b) The contribution is based upon previous work that, to the best
  of my knowledge, is covered under an appropriate open source
  license and I have the right under that license to submit that
  work with modifications, whether created in whole or in part
  by me, under the same open source license (unless I am
  permitted to submit under a different license), as indicated
  in the file; or

* (c) The contribution was provided directly to me by some other
  person who certified (a), (b) or (c) and I have not modified
  it.

* (d) I understand and agree that this project and the contribution
  are public and that a record of the contribution (including all
  personal information I submit with it, including my sign-off) is
  maintained indefinitely and may be redistributed consistent with
  this project or the open source license(s) involved.
-->

#### Checklist

- [ ] run `npm run test` and `npm run benchmark`
- [ ] tests and/or benchmarks are included
- [ ] documentation is changed or added
- [ ] commit message and code follows the [Developer's Certification of Origin](https://github.com/fastify/.github/blob/master/CONTRIBUTING.md#developers-certificate-of-origin-11)
      and the [Code of conduct](https://github.com/fastify/.github/blob/master/CODE_OF_CONDUCT.md)