@Uzlopak 2023-01-24
For now I think it is more important to decide if we want a strict system.